sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "akasha-client"
version = "1.0.0"
description = "Python SDK for Akasha — The Shared Cognitive Fabric for Intelligent Agent Systems"
readme = "README.md"
license = {text = "ASL-1.0"}
requires-python = ">=3.10"
authors = [{name = "Alejandro Sánchez Losa", email = "dev@alejandrosl.com"}]
keywords = ["akasha", "shared-memory", "agents", "grpc", "real-time"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "msgpack>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/ocuil/akasha-public"
Repository = "https://github.com/ocuil/akasha-public"

[tool.hatch.build.targets.wheel]
packages = ["src/akasha"]

[tool.ruff]
line-length = 100
target-version = "py310"
